adjust my patches route
