+++ src/components/Content/index.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import { Route, Redirect, Switch } from "react-router-dom";
 import { Task } from "pages/Task";
 import { Patch } from "pages/Patch";
-import { MyPatches } from "pages/MyPatches";
+import { UserPatches } from "pages/UserPatches";
 import { Login } from "pages/Login";
 import { CommitQueue } from "pages/CommitQueue";
 import { PrivateRoute } from "components/PrivateRoute";
 import { Navbar } from "components/Navbar";
-import { routes, getUserPatchesRoute } from "constants/routes";
+import { routes } from "constants/routes";
 import { FullPageLoad } from "components/Loading/FullPageLoad";
 import { useAuthStateContext } from "context/auth";
 import { useQuery } from "@apollo/react-hooks";
@@ -17,6 +17,7 @@ import { PageLayout } from "components/styles/Layout";
 import { PageDoesNotExist } from "pages/404";
 import { ConfigurePatch } from "pages/ConfigurePatch";
 import { Preferences } from "pages/Preferences";
+import { MyPatches } from "pages/MyPatches";
 
 export const Content: React.FC = () => {
   const { isAuthenticated, initialLoad } = useAuthStateContext();
@@ -24,9 +25,9 @@ export const Content: React.FC = () => {
   // this top-level query is required for authentication to work
   // afterware is used at apollo link level to authenticate or deauthenticate user based on response to query
   // therefore this could be any query as long as it is top-level
-  const { data, loading } = useQuery<GetUserQuery>(GET_USER);
+  useQuery<GetUserQuery>(GET_USER);
 
-  if (loading || (!isAuthenticated && initialLoad)) {
+  if (!isAuthenticated && initialLoad) {
     return <FullPageLoad />;
   }
 
@@ -37,10 +38,8 @@ export const Content: React.FC = () => {
         <PrivateRoute path={routes.task} component={Task} />
         <PrivateRoute path={routes.configurePatch} component={ConfigurePatch} />
         <PrivateRoute path={routes.patch} component={Patch} />
-        <PrivateRoute exact path={routes.myPatches}>
-          <Redirect to={getUserPatchesRoute(data.user.userId)} />
-        </PrivateRoute>
-        <PrivateRoute path={routes.userPatches} component={MyPatches} />
+        <PrivateRoute exact path={routes.myPatches} component={MyPatches} />
+        <PrivateRoute path={routes.userPatches} component={UserPatches} />
         <PrivateRoute path={routes.commitQueue} component={CommitQueue} />
         <PrivateRoute path={routes.preferences} component={Preferences} />
         <PrivateRoute exact path="/">
